Mengganti nama package dan menambahkan display name
