+++ app/google-services.json
@@ -1,41 +1,70 @@
 {
   "project_info": {
-    "project_number": "347343223443",
-    "project_id": "test-finalproject",
-    "storage_bucket": "test-finalproject.appspot.com"
+    "project_number": "427927402673",
+    "project_id": "pristine-sphere-343506",
+    "storage_bucket": "pristine-sphere-343506.appspot.com"
   },
   "client": [
     {
       "client_info": {
-        "mobilesdk_app_id": "1:347343223443:android:7ca9cc856fcd269133ab66",
+        "mobilesdk_app_id": "1:427927402673:android:f2e774d791423265ec0a90",
         "android_client_info": {
-          "package_name": "com.example.pneumoniadetector"
+          "package_name": "com.bangkit.pneumoniadetector"
         }
       },
       "oauth_client": [
         {
-          "client_id": "347343223443-90ts5uug54bqq6ocsjtdm7acejbpj430.apps.googleusercontent.com",
+          "client_id": "427927402673-r77fv0non0kskau47ab91iarrc8jt2n4.apps.googleusercontent.com",
           "client_type": 1,
           "android_info": {
-            "package_name": "com.example.pneumoniadetector",
+            "package_name": "com.bangkit.pneumoniadetector",
             "certificate_hash": "0061fc001d929288278fdbcf341a7d3c80fad5c3"
           }
         },
         {
-          "client_id": "347343223443-9u9je4uv9mln6pt6neaogvp9a26595eq.apps.googleusercontent.com",
+          "client_id": "427927402673-6rfs79lu7ffgnkotrcd5bq9h59apntjv.apps.googleusercontent.com",
           "client_type": 3
         }
       ],
       "api_key": [
         {
-          "current_key": "AIzaSyC7ylpltcSF_kNkNWc9ccCLDEpe0pEae10"
+          "current_key": "AIzaSyCbkS7rKgns1JVg1NBdYUpxlf8pGCj9WEQ"
         }
       ],
       "services": {
         "appinvite_service": {
           "other_platform_oauth_client": [
             {
-              "client_id": "347343223443-9u9je4uv9mln6pt6neaogvp9a26595eq.apps.googleusercontent.com",
+              "client_id": "427927402673-6rfs79lu7ffgnkotrcd5bq9h59apntjv.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:427927402673:android:185560ea2e9883c3ec0a90",
+        "android_client_info": {
+          "package_name": "com.example.pneumoniadetector"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "427927402673-6rfs79lu7ffgnkotrcd5bq9h59apntjv.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyCbkS7rKgns1JVg1NBdYUpxlf8pGCj9WEQ"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "427927402673-6rfs79lu7ffgnkotrcd5bq9h59apntjv.apps.googleusercontent.com",
               "client_type": 3
             }
           ]
