+++ app/src/main/java/com/bangkit/pneumoniadetector/data/remote/response/Response.kt
@@ -1,0 +1,2 @@
+package com.bangkit.pneumoniadetector.data.remote.response
+
