+++ app/src/main/java/com/bangkit/pneumoniadetector/data/remote/retrofit/ApiService.kt
@@ -1,0 +1,4 @@
+package com.bangkit.pneumoniadetector.data.remote.retrofit
+
+interface ApiService {
+}
