+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/MainViewModel.kt
@@ -1,0 +1,7 @@
+package com.bangkit.pneumoniadetector.ui
+
+import androidx.lifecycle.ViewModel
+
+class MainViewModel : ViewModel() {
+
+}
