+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/ViewModelFactory.kt
@@ -1,0 +1,27 @@
+package com.bangkit.pneumoniadetector.ui
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class ViewModelFactory :  ViewModelProvider.NewInstanceFactory(){
+
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        return when {
+            modelClass.isAssignableFrom(MainViewModel::class.java) -> {
+                MainViewModel() as T
+            }
+            else -> throw IllegalArgumentException("Unknown ViewModel class: " + modelClass.name)
+        }
+    }
+
+    companion object {
+        @Volatile
+        private var instance: ViewModelFactory? = null
+        fun getInstance(): ViewModelFactory =
+            instance ?: synchronized(this) {
+                instance ?: ViewModelFactory()
+            }.also { instance = it }
+    }
+
+}
