+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/detail/EditDetailActivity.kt
@@ -1,0 +1,12 @@
+package com.bangkit.pneumoniadetector.ui.detail
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.bangkit.pneumoniadetector.R
+
+class EditDetailActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_edit_detail)
+    }
+}
