+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/history/HistoryAdapter.kt
@@ -1,0 +1,4 @@
+package com.bangkit.pneumoniadetector.ui.history
+
+class HistoryAdapter {
+}
