+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/history/HistoryFragment.kt
@@ -1,0 +1,40 @@
+package com.bangkit.pneumoniadetector.ui.history
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.bangkit.pneumoniadetector.databinding.FragmentHistoryBinding
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+
+class HistoryFragment : Fragment() {
+
+    private var _binding: FragmentHistoryBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        //val historyViewModel = ViewModelProvider(this)[HistoryViewModel::class.java]
+
+        _binding = FragmentHistoryBinding.inflate(inflater, container, false)
+
+        val user = Firebase.auth.currentUser
+
+        binding.textViewTitle.text = "Hi, " + user?.displayName.toString()
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
