+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/history/HistoryViewModel.kt
@@ -1,0 +1,13 @@
+package com.bangkit.pneumoniadetector.ui.history
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class HistoryViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = ""
+    }
+    val text: LiveData<String> = _text
+}
