+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/home/HomeFragment.kt
@@ -1,0 +1,80 @@
+package com.bangkit.pneumoniadetector.ui.home
+
+import android.Manifest
+import android.content.Intent
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.core.app.ActivityCompat
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.bangkit.pneumoniadetector.databinding.FragmentHomeBinding
+import com.bangkit.pneumoniadetector.tools.GeneralTools
+import com.bangkit.pneumoniadetector.ui.camera.CameraActivity
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+
+class HomeFragment : Fragment() {
+
+    private var _binding: FragmentHomeBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val homeViewModel =
+            ViewModelProvider(this)[HomeViewModel::class.java]
+
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+
+        val user = Firebase.auth.currentUser
+
+        binding.textViewTitle.text = "Hi, " + user?.displayName.toString()
+        //
+//        val textView: TextView = binding.textHome
+//        homeViewModel.text.observe(viewLifecycleOwner) {
+//            textView.text = it
+//        }
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        // when click cardView, goes to CameraActivity when having camera permission granted
+        binding.materialCardView.setOnClickListener { goesToCameraActivity() }
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    // Method for clicking materialCardView that will goes to CameraActivity when permission granted
+    fun goesToCameraActivity() {
+
+        // check camera permission
+        if(!GeneralTools.allPermissionGranted(REQUIRED_CAMERA_PERMISSION, requireContext())){
+            ActivityCompat.requestPermissions(
+                requireActivity(),
+                REQUIRED_CAMERA_PERMISSION,
+                REQUEST_CAMERA_CODE_PERMISSION
+            )
+        }
+        else{
+            val intent = Intent(requireContext(), CameraActivity::class.java)
+            startActivity(intent)
+        }
+    }
+
+    companion object{
+        val REQUIRED_CAMERA_PERMISSION = arrayOf(Manifest.permission.CAMERA)
+        const val REQUEST_CAMERA_CODE_PERMISSION = 10
+    }
+}
