+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/home/RecentAdapter.kt
@@ -1,0 +1,4 @@
+package com.bangkit.pneumoniadetector.ui.home
+
+class RecentAdapter {
+}
