+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/login/LoginActivity.kt
@@ -1,0 +1,87 @@
+package com.bangkit.pneumoniadetector.ui.login
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.util.Log
+import android.widget.Toast
+import com.bangkit.pneumoniadetector.databinding.ActivityLoginBinding
+import com.bangkit.pneumoniadetector.ui.MainActivity
+import com.bangkit.pneumoniadetector.ui.register.RegisterActivity
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+
+class LoginActivity : AppCompatActivity() {
+
+    private lateinit var auth: FirebaseAuth
+    private lateinit var binding: ActivityLoginBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = ActivityLoginBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        // Initialize Firebase Auth
+        auth = Firebase.auth
+
+        setupAction()
+    }
+
+    private fun setupAction() {
+        binding.btnLogin.setOnClickListener {
+            val email = binding.etEmail.text.toString()
+            val password = binding.etPassword.text.toString()
+            when {
+                email.isEmpty() -> {
+                    binding.etEmail.error = "Masukkan email"
+                }
+                password.isEmpty() -> {
+                    binding.etPassword.error = "Masukkan password"
+                }
+                else -> {
+                    auth.signInWithEmailAndPassword(email, password)
+                        .addOnCompleteListener(this) { task ->
+                            if (task.isSuccessful) {
+                                // Sign in success, update UI with the signed-in user's information
+                                Log.d(TAG, "signInWithEmail:success")
+                                //val user = auth.currentUser
+                                //updateUI(user)
+                                val intent = Intent(this, MainActivity::class.java)
+                                startActivity(intent)
+                                finish()
+                            } else {
+                                // If sign in fails, display a message to the user.
+                                Log.w(TAG, "signInWithEmail:failure", task.exception)
+                                Toast.makeText(baseContext, "Authentication failed.",
+                                    Toast.LENGTH_SHORT).show()
+                                //updateUI(null)
+                            }
+                        }
+                }
+            }
+        }
+
+        binding.tvSignUpHere.setOnClickListener {
+            val intent = Intent(this, RegisterActivity::class.java)
+            startActivity(intent)
+            finish()
+        }
+    }
+
+    public override fun onStart() {
+        super.onStart()
+        // Check if user is signed in (non-null) and update UI accordingly.
+        val currentUser = Firebase.auth.currentUser
+        if(currentUser != null){
+            val intent = Intent(this, MainActivity::class.java)
+            startActivity(intent)
+            finish()
+        }
+    }
+
+    companion object {
+        private const val TAG = "LoginActivity"
+    }
+}
