+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/profile/ProfileFragment.kt
@@ -1,0 +1,63 @@
+package com.bangkit.pneumoniadetector.ui.profile
+
+import android.content.Intent
+import android.os.Bundle
+import android.util.Log
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.bangkit.pneumoniadetector.databinding.FragmentProfileBinding
+import com.bangkit.pneumoniadetector.ui.login.LoginActivity
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+
+class ProfileFragment : Fragment() {
+
+    private var _binding: FragmentProfileBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val profileViewModel =
+            ViewModelProvider(this)[ProfileViewModel::class.java]
+
+        _binding = FragmentProfileBinding.inflate(inflater, container, false)
+
+        val user = Firebase.auth.currentUser
+
+        binding.textViewName.text = user?.displayName.toString()
+        binding.textViewEmail.text = user?.email.toString()
+
+        setupAction()
+
+        return binding.root
+    }
+
+    private fun setupAction() {
+        binding.btnLogout.setOnClickListener {
+            Log.e(TAG, "setupAction: tes", )
+            Firebase.auth.signOut()
+
+            val intent = Intent(activity, LoginActivity::class.java)
+            startActivity(intent)
+            activity?.finish()
+        }
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    companion object {
+        private const val TAG = "ProfileFragment"
+    }
+}
