+++ app/src/main/java/com/bangkit/pneumoniadetector/ui/profile/ProfileViewModel.kt
@@ -1,0 +1,13 @@
+package com.bangkit.pneumoniadetector.ui.profile
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class ProfileViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is notifications Fragment"
+    }
+    val text: LiveData<String> = _text
+}
