--- app/src/main/java/com/example/pneumoniadetector/data/remote/response/Response.kt
@@ -1,2 +1,0 @@
-package com.example.pneumoniadetector.data.remote.response
-
