--- app/src/main/java/com/example/pneumoniadetector/data/remote/retrofit/ApiService.kt
@@ -1,4 +1,0 @@
-package com.example.pneumoniadetector.data.remote.retrofit
-
-interface ApiService {
-}
