--- app/src/main/java/com/example/pneumoniadetector/ui/MainViewModel.kt
@@ -1,12 +1,0 @@
-package com.example.pneumoniadetector.ui
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import com.google.firebase.auth.FirebaseUser
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
-
-class MainViewModel : ViewModel() {
-
-}
