--- app/src/main/java/com/example/pneumoniadetector/ui/ViewModelFactory.kt
@@ -1,30 +1,0 @@
-package com.example.pneumoniadetector.ui
-
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-
-class ViewModelFactory :  ViewModelProvider.NewInstanceFactory(){
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return when {
-            modelClass.isAssignableFrom(MainViewModel::class.java) -> {
-                MainViewModel() as T
-            }
-            else -> throw IllegalArgumentException("Unknown ViewModel class: " + modelClass.name)
-        }
-    }
-
-    companion object {
-        @Volatile
-        private var instance: ViewModelFactory? = null
-        fun getInstance(): ViewModelFactory =
-            instance ?: synchronized(this) {
-                instance ?: ViewModelFactory()
-            }.also { instance = it }
-    }
-
-}
