--- app/src/main/java/com/example/pneumoniadetector/ui/detail/EditDetailActivity.kt
@@ -1,12 +1,0 @@
-package com.example.pneumoniadetector.ui.detail
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.pneumoniadetector.R
-
-class EditDetailActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_edit_detail)
-    }
-}
