--- app/src/main/java/com/example/pneumoniadetector/ui/history/HistoryAdapter.kt
@@ -1,4 +1,0 @@
-package com.example.pneumoniadetector.ui.history
-
-class HistoryAdapter {
-}
