--- app/src/main/java/com/example/pneumoniadetector/ui/history/HistoryFragment.kt
@@ -1,35 +1,0 @@
-package com.example.pneumoniadetector.ui.history
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
-import com.example.pneumoniadetector.databinding.FragmentHistoryBinding
-
-class HistoryFragment : Fragment() {
-
-    private var _binding: FragmentHistoryBinding? = null
-
-    // This property is only valid between onCreateView and
-    // onDestroyView.
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        //val historyViewModel = ViewModelProvider(this)[HistoryViewModel::class.java]
-
-        _binding = FragmentHistoryBinding.inflate(inflater, container, false)
-
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
