--- app/src/main/java/com/example/pneumoniadetector/ui/history/HistoryViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.pneumoniadetector.ui.history
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class HistoryViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = ""
-    }
-    val text: LiveData<String> = _text
-}
