--- app/src/main/java/com/example/pneumoniadetector/ui/home/RecentAdapter.kt
@@ -1,4 +1,0 @@
-package com.example.pneumoniadetector.ui.home
-
-class RecentAdapter {
-}
