--- app/src/main/java/com/example/pneumoniadetector/ui/profile/EditProfileActivity.kt
@@ -1,12 +1,0 @@
-package com.example.pneumoniadetector.ui.profile
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.pneumoniadetector.R
-
-class EditProfileActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_edit_profile)
-    }
-}
