--- app/src/main/java/com/example/pneumoniadetector/ui/profile/ProfileViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.pneumoniadetector.ui.profile
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ProfileViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
-    }
-    val text: LiveData<String> = _text
-}
