--- app/src/main/java/com/example/pneumoniadetector/ui/register/RegisterActivity.kt
@@ -1,109 +1,0 @@
-package com.example.pneumoniadetector.ui.register
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.util.Log
-import android.widget.Toast
-import com.example.pneumoniadetector.R
-import com.example.pneumoniadetector.databinding.ActivityLoginBinding
-import com.example.pneumoniadetector.databinding.ActivityMainBinding
-import com.example.pneumoniadetector.databinding.ActivityRegisterBinding
-import com.example.pneumoniadetector.ui.MainActivity
-import com.example.pneumoniadetector.ui.login.LoginActivity
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.auth.ktx.userProfileChangeRequest
-import com.google.firebase.ktx.Firebase
-
-class RegisterActivity : AppCompatActivity() {
-
-    private lateinit var auth: FirebaseAuth
-    private lateinit var binding: ActivityRegisterBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = ActivityRegisterBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        // Initialize Firebase Auth
-        auth = Firebase.auth
-
-        setupAction()
-    }
-
-    public override fun onStart() {
-        super.onStart()
-        // Check if user is signed in (non-null) and update UI accordingly.
-        val currentUser = Firebase.auth.currentUser
-        if(currentUser != null){
-            val intent = Intent(this, MainActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
-    }
-
-    private fun setupAction() {
-        binding.btnRegister.setOnClickListener {
-            val name = binding.etName.text.toString()
-            val email = binding.etEmail.text.toString()
-            val password = binding.etPassword.text.toString()
-            when {
-                name.isEmpty() -> {
-                    binding.etName.error = "Masukkan email"
-                }
-                email.isEmpty() -> {
-                    binding.etEmail.error = "Masukkan email"
-                }
-                password.isEmpty() -> {
-                    binding.etPassword.error = "Masukkan password"
-                }
-                else -> {
-                    auth.createUserWithEmailAndPassword(email, password)
-                        .addOnCompleteListener(this) { task ->
-                            if (task.isSuccessful) {
-                                // Sign in success, update UI with the signed-in user's information
-                                Log.d(TAG, "createUserWithEmail:success")
-                                Toast.makeText(baseContext, "Berhasil mendaftarkan akun.",
-                                    Toast.LENGTH_SHORT).show()
-
-                                val user = auth.currentUser
-                                val profileUpdates = userProfileChangeRequest {
-                                    displayName = name
-                                }
-
-                                user!!.updateProfile(profileUpdates)
-                                    .addOnCompleteListener { task ->
-                                        if (task.isSuccessful) {
-                                            Log.d(TAG, "User profile updated.")
-                                        }
-                                    }
-
-                                val intent = Intent(this, MainActivity::class.java)
-                                startActivity(intent)
-                                finish()
-
-                            } else {
-                                // If sign in fails, display a message to the user.
-                                Log.w(TAG, "createUserWithEmail:failure", task.exception)
-                                Toast.makeText(baseContext, "Authentication failed.",
-                                    Toast.LENGTH_SHORT).show()
-                                //updateUI(null)
-                            }
-                        }
-                }
-            }
-        }
-
-        binding.tvLogin2.setOnClickListener {
-            startActivity(Intent(this, LoginActivity::class.java))
-            finish()
-        }
-    }
-
-    companion object {
-        private const val TAG = "RegisterActivity"
-    }
-
-}
