+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.example.pneumoniadetector.ui.home.HomeFragment"
+        android:name="com.bangkit.pneumoniadetector.ui.home.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/navigation_history"
-        android:name="com.example.pneumoniadetector.ui.history.HistoryFragment"
+        android:name="com.bangkit.pneumoniadetector.ui.history.HistoryFragment"
         android:label="@string/title_history"
         tools:layout="@layout/fragment_history" />
 
     <fragment
         android:id="@+id/navigation_profile"
-        android:name="com.example.pneumoniadetector.ui.profile.ProfileFragment"
+        android:name="com.bangkit.pneumoniadetector.ui.profile.ProfileFragment"
         android:label="@string/title_profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
